move fd util to system.py
